requirements.txt
flask==2.3.2
requests==2.31.0
Flask-Alembic==2.0.1
psycopg2==2.9.6
python-dotenv==1.0.0
